+++ app/(protected)/(with-layout)/billing/page.tsx
@@ -1,0 +1,7 @@
+export default function Page() {
+  return (
+    <>
+      <h1>billing Page Content</h1>
+    </>
+  );
+}
