+++ app/(protected)/(with-layout)/layout.tsx
@@ -1,0 +1,9 @@
+import Sidebar from "@/components/auth/sidebar";
+export default function SidebarLayout({ children }) {
+  return (
+    <div className="flex">
+      <Sidebar />
+      <div className="flex">{children}</div>
+    </div>
+  );
+}
