+++ app/(protected)/(with-layout)/profile/page.tsx
@@ -1,0 +1,7 @@
+export default function Page() {
+  return (
+    <>
+      <h1>profile Page Content</h1>
+    </>
+  );
+}
