+++ app/(protected)/(with-layout)/settings/page.tsx
@@ -1,0 +1,4 @@
+const SettingPage = async () => {
+  return <>SettingPage Content</>;
+};
+export default SettingPage;
