+++ app/(protected)/(with-layout)/tokens/page.tsx
@@ -1,0 +1,8 @@
+export default function Page() {
+  return (
+    <>
+      <h1>API keys Page Content</h1>
+    </>
+
+  );
+}
