--- app/(protected)/billing/page.tsx
@@ -1,14 +1,0 @@
-import SidebarLayout from "@/components/auth/sidebar";
-
-export default function Page() {
-  return (
-    <>
-      <SidebarLayout >
-        <div>
-          <h1>billing Page Content</h1>
-        </div>
-      </SidebarLayout>
-    </>
-
-  );
-}
