--- app/(protected)/profile/page.tsx
@@ -1,15 +1,0 @@
-import SidebarLayout from "@/components/auth/sidebar";
-
-export default function Page() {
-  return (
-    <>
-      <SidebarLayout >
-        <div>
-          <h1>profile Page Content</h1>
-        </div>
-      </SidebarLayout>
-
-    </>
-
-  );
-}
