--- app/(protected)/settings/page.tsx
@@ -1,7 +1,0 @@
-import { Button } from "@/components/ui/button";
-
-const SettingPage = async () => {
-  return <div>SettingPage Content</div>;
-};
-
-export default SettingPage;
